Prevent line breaks in Misc text

+++ encryption.c
@@ -151,7 +151,7 @@ void updateBuffer(void)
 {
 	buffer = NULL;
 	if (state.entryCount > 0)
-		buffer = (char *)malloc(state.entryCount * MAXLINE);
+		buffer = (char *)malloc((size_t)state.entryCount * MAXLINE);
 	else
 		buffer = (char *)malloc(16);
